Rename book form component
